W8-A3/requirements.txt
langchain-core==1.2.9
langchain-community==0.4.1
langchain-openai==1.1.7
pypdf==6.6.2
python-dotenv==1.2.1
